✨ [#17] feat : 프로젝트/스터디 컴포넌트 추가 및 프로젝트 더미 데이터 생성
