+++ dummy/ProjectsData.ts
@@ -1,0 +1,132 @@
+export interface ProjectDummyData {
+  id: number;
+  title: string;
+  categories: string[];
+  state: boolean;
+  totalUser: number;
+  currentUser: number;
+  like: number;
+  imgUrl: string;
+}
+export const ProjectsData: ProjectDummyData[] = [
+  {
+    id: 1,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 4,
+    currentUser: 2,
+    like: 20,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 2,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Next'],
+    state: true,
+    totalUser: 5,
+    currentUser: 1,
+    like: 10,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 3,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 6,
+    currentUser: 2,
+    like: 4,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 4,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 3,
+    currentUser: 1,
+    like: 4,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 5,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 6,
+    currentUser: 2,
+    like: 4,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 6,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 6,
+    currentUser: 2,
+    like: 6,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 7,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 6,
+    currentUser: 2,
+    like: 30,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 8,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 4,
+    currentUser: 3,
+    like: 7,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 7,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 6,
+    currentUser: 2,
+    like: 30,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 8,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 4,
+    currentUser: 3,
+    like: 7,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 9,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 10,
+    currentUser: 2,
+    like: 10,
+    imgUrl: '/images/dummy_image.png',
+  },
+  {
+    id: 10,
+    title: 'React + Spring 스터디 / 프로젝트 모집 사이트 같이 만드실분',
+    categories: ['React', 'Typescript', 'MySQL', 'NodeJs', 'Spring'],
+    state: true,
+    totalUser: 8,
+    currentUser: 5,
+    like: 7,
+    imgUrl: '/images/dummy_image.png',
+  },
+];
