+++ pages/index.tsx
@@ -2,16 +2,6 @@ import { NextPage } from 'next';
 import MainNewContent from '../components/main/MainNewContent';
 import Layout from '../components/Layout';
 
-export interface ProjectDummyData {
-  title: string;
-  categories: string[];
-  state: boolean;
-  totalUser: number;
-  currentUser: number;
-  like: number;
-  imgUrl: string;
-}
-
 const Home: NextPage = () => {
   return (
     <Layout>
