+++ pages/study/index.tsx
@@ -1,7 +1,18 @@
 import { NextPage } from 'next';
+import FilteredBox from '../../components/common/FilteredBox';
+import Layout from '../../components/Layout';
+import ProjectStudyHeader from '../../components/project-study/header';
+import { FilterBoxWrapper } from '../projects/styled';
 
 const StudyPage: NextPage = () => {
-  return <div></div>;
+  return (
+    <Layout>
+      <ProjectStudyHeader />
+      <FilterBoxWrapper>
+        <FilteredBox />
+      </FilterBoxWrapper>
+    </Layout>
+  );
 };
 
 export default StudyPage;
